Refactor helper classes and MainViewModel
